requirements.txt
atproto==0.0.65
autobahn==25.12.2
automat==25.4.16
boto3==1.42.97
botocore==1.42.97
cbor2==6.0.1
celery[redis]==5.6.3
cfgv==3.5.0
channels==4.3.2
channels-redis==4.3.0
constantly==23.10.4
coverage==7.13.5
cryptography==46.0.7
cssbeautifier==1.15.4
daphne==4.2.1
Deprecated==1.3.1
dj-database-url==3.1.2
dj-rest-auth==7.2.0
django-allauth==65.16.1
django-anymail==15.0
django-cors-headers==4.9.0
django-import-export==4.4.0
django-storages==1.14.6
django-stubs-ext==6.0.3
django-stubs==6.0.3
django-unfold==0.91.0
Django==6.0.4
djangochannelsrestframework==1.3.0
djangorestframework-simplejwt==5.5.1
djangorestframework==3.17.1
djlint==1.36.4
dnspython==2.8.0
docutils==0.22.4
drf-nested-routers==0.95.0
drf-spectacular==0.29.0
drf-standardized-errors==0.15.0
editorconfig==0.17.1
feedparser==6.0.12
gunicorn==25.3.0
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.19
incremental==24.11.0
inflection==0.5.1
jmespath==1.1.0
jsbeautifier==1.15.4
langchain-core==1.3.2
langchain-protocol==0.0.15
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.12
langgraph-sdk==0.3.13
langgraph==1.1.10
langsmith==0.7.37
libipld==3.3.2
librt==0.9.0
markdown==3.10.2
Mastodon.py==2.2.1
msgpack==1.1.2
mypy==1.20.2
nodeenv==1.10.0
opentelemetry-api>=1.41.1
opentelemetry-exporter-otlp-proto-http>=1.28,<2
opentelemetry-instrumentation-celery>=0.49b0,<1
opentelemetry-instrumentation-django>=0.49b0,<1
opentelemetry-sdk>=1.41.1
ormsgpack==1.12.2
pillow==12.2.0
praw==7.8.1
pre-commit==4.6.0
psycopg[binary]==3.3.3
py-ubjson==0.16.1
pyasn1-modules==0.4.2
pyasn1==0.6.3
pylint-django==2.7.0
pylint-plugin-utils==0.9.0
pyopenssl==26.1.0
pytest-django==4.12.0
pytest-mock==3.15.1
pytest==9.0.3
python-dotenv==1.2.2
qdrant-client==1.17.1
requests-toolbelt==1.0.0
ruff==0.15.12
s3transfer==0.16.1
sentence-transformers==5.4.1
service-identity==24.2.0
standardwebhooks==1.0.1
structlog==25.5.0
svix==1.92.2
twisted==25.5.0
txaio==25.12.2
types-Deprecated==1.3.1.20260408
types-python-dateutil==2.9.0.20260408
types-pyyaml==6.0.12
types-requests==2.32.4.20250913
ujson==5.12.0
uritemplate==4.2.0
uuid-utils==0.14.1
watchdog==6.0.0
websockets==15.0.1
wrapt==2.1.2
xxhash==3.7.0
zope-interface==8.4
